+++ components/CTAButton/CTAButton.js
@@ -3,21 +3,28 @@ import PropTypes from 'prop-types';
 
 import Link from '../Link';
 
+import css from './CTAButton.module.scss';
+
 class CTAButton extends Component {
   static propTypes = {
     href: PropTypes.string.isRequired,
     size: PropTypes.string,
     target: PropTypes.string,
+    className: PropTypes.string,
   };
 
   render() {
-    const { children, href, size, target } = this.props;
+    const { children, href, size, target, className } = this.props;
 
     // TODO: Make this DRY
     return href && href.match('^http') ? (
       <a
         href={href}
-        className={`button ${size ? `button--${size}` : ''}`}
+        className={[
+          css.ctaButton,
+          size ? css.ctaButtonSm : '',
+          className || '',
+        ].join(' ')}
         target={target}
       >
         {children}
@@ -26,7 +33,11 @@ class CTAButton extends Component {
       <Link as={href}>
         <a
           // href={href}
-          className={`button ${size ? `button--${size}` : ''}`}
+          className={[
+            css.ctaButton,
+            size ? css.ctaButtonSm : '',
+            className || '',
+          ].join(' ')}
           target={target}
         >
           {children}