+++ components/TileButtonGroup/TileButtonGroup.js
@@ -2,7 +2,8 @@ import { Component } from 'react';
 import PropTypes from 'prop-types';
 
 import CTAButton from '../CTAButton';
-// import './TileButtonGroup.css';
+
+import css from './TileButtonGroup.module.scss';
 
 class TileButtonGroup extends Component {
   static propTypes = {
@@ -31,21 +32,35 @@ class TileButtonGroup extends Component {
     } = this.props;
 
     return (
-      <div className="tile-button-group tile-button-group--alt">
+      <div className={css.tileButtonGroup}>
         {primaryUrl && (
-          <CTAButton href={primaryUrl} target={primaryTarget}>
+          <CTAButton
+            href={primaryUrl}
+            target={primaryTarget}
+            className={css.primaryButton}
+          >
             {primaryText}
           </CTAButton>
         )}
 
         {secondaryUrl && (
-          <CTAButton href={secondaryUrl} target={secondaryTarget} size="sm">
+          <CTAButton
+            href={secondaryUrl}
+            target={secondaryTarget}
+            size="sm"
+            className={css.secondaryButton}
+          >
             {secondaryText}
           </CTAButton>
         )}
 
         {tertiaryUrl && (
-          <CTAButton href={tertiaryUrl} target={tertiaryTarget} size="sm">
+          <CTAButton
+            href={tertiaryUrl}
+            target={tertiaryTarget}
+            size="sm"
+            className={css.secondaryButton}
+          >
             {tertiaryText}
           </CTAButton>
         )}