+++ CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.8.2] - 2020-04-01
+## [1.8.2] - 2020-04-XX
 
-- Add `DXLabLogo` and `SLNSWLogo` components
+- Add `DXLabLogo`, `SLNSWLogo` components
+- Refactor `Button` to `CTAButton`
+- Recreate `Button` component to be more generic
+- Add `MenuIconButton`
 
 ## [1.8.1] - 2020-04-01